requirements.txt
# Web框架
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
python-dotenv==1.0.1

# 数据库
sqlalchemy==2.0.28
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.2
pymilvus==2.3.6

# AI & ML
openai==1.14.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
transformers==4.38.2
sentence-transformers==2.5.1

# 工具
pydantic==2.6.3
pydantic-settings==2.2.1
tenacity==8.2.3
httpx==0.27.0
aiohttp==3.9.3
python-dateutil==2.9.0.post0
pytz==2024.1

# 监控
prometheus-client==0.20.0
structlog==24.1.0

# 测试
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0
aioresponses==0.7.6 